requirements.txt
Django==3.2.5
djangorestframework==3.12.4
asgiref==3.8.1
automium==0.2.6
automium_web==0.1.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
legacy==0.1.7
MarkupSafe==2.1.5
sqlparse==0.5.0
tzdata==2024.1
Werkzeug==3.0.3
djangorestframework==3.12.4
pip~=21.2.4
wheel~=0.37.0
pytz~=2024.1
requests~=2.32.2
setuptools~=58.0.4
zipp~=3.18.2
future~=0.18.2